refactor: move schemas out of the shared folder and into their own files
