+++ src/commonmark/editor.ts
@@ -12,7 +12,7 @@ import {
     editableCheck,
     readonlyPlugin,
 } from "../shared/prosemirror-plugins/readonly";
-import { CodeStringParser, commonmarkSchema } from "../shared/schema";
+import { CodeStringParser } from "../shared/schema";
 import { deepMerge } from "../shared/utils";
 import {
     BaseView,
@@ -21,6 +21,7 @@ import {
 } from "../shared/view";
 import { createMenu } from "./commands";
 import { allKeymaps } from "./key-bindings";
+import { commonmarkSchema } from "./schema";
 
 export type CommonmarkOptions = CommonViewOptions;
